Move from __tests__ to src/__tests__
