--- __tests__/fixtures/index.ts
@@ -1,1 +1,0 @@
-export { helper } from '../../src';
