+++ src/__tests__/fixtures/basic/declaration/input.js
