+++ src/__tests__/fixtures/basic/declaration/output.js
