+++ src/__tests__/fixtures/basic/dependencies/input.js
