+++ src/__tests__/fixtures/basic/dependencies/output.js
