+++ src/__tests__/fixtures/basic/officialMix/input.js
