+++ src/__tests__/fixtures/basic/rename/input.js
