+++ src/__tests__/fixtures/basic/rename/output.js
