+++ src/__tests__/fixtures/error/babelAlreadyDefined/input.js
