+++ src/__tests__/fixtures/error/doNotExportDefault/input.js
