+++ src/__tests__/fixtures/error/exportAll/input.js
