+++ src/__tests__/fixtures/error/exportNamed/input.js
